add a superclass for DB workers, for avoiding code repetition
